additional logging added; warnings removed;

+++ ios/UnityAds/UnityAdsZone/UnityAdsZone.m
@@ -77,7 +77,7 @@
   [self._options setObject:stringValue forKey:kUnityAdsZoneNoOfferScreenKey];
 }
 
-- (int)allowVideoSkipInSeconds {
+- (NSInteger)allowVideoSkipInSeconds {
   return [[self._options valueForKey:kUnityAdsZoneAllowVideoSkipInSecondsKey] integerValue];
 }
